refactor(interface): reorganize socket interfaces
move all socket-related types into server directory.
