+++ client/src/types/socket.ts
@@ -2,7 +2,7 @@ import type { Socket } from "socket.io-client";
 import type {
   ClientToServerEvents,
   ServerToClientEvents
-} from "@socket/types";
+} from "@server";
 
 /** Shared socket from server side code are configured here globally */
 export type BiseoSocket = Socket<
