+++ client/tsconfig.json
@@ -5,7 +5,8 @@
     "baseUrl": "src",
     "paths": {
       "@/*": ["*"],
-      "@socket/*": ["../../socket/*"]
+      "@server": ["../../server/src/interface"],
+      "@server/*": ["../../server/src/interface/*"]
     },
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
     "allowJs": false,
