+++ server/src/interface/helpers.ts
