--- server/src/model/index.ts
@@ -1,4 +1,0 @@
-export type {
-  ChatServerToClientEvents,
-  ChatClientToServerEvents,
-} from "./chat";
